requirements.txt
Flask
Flask-PyMongo
flask-jwt-extended
PyJWT
Flask-Bcrypt
Flask-Cors
Flask-JWT
pymongo[tls,srv]==3.6.1